[#2] feature : 아이템 Wishlist 관련 기능

+++ src/main/java/com/sparta/ecommerce/domain/iteminfo/repository/ItemInfoRepository.java
@@ -3,5 +3,5 @@ package com.sparta.ecommerce.domain.iteminfo.repository;
 import com.sparta.ecommerce.domain.iteminfo.entity.ItemInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemInfoRepository extends JpaRepository<ItemInfo, Integer> {
+public interface ItemInfoRepository extends JpaRepository<ItemInfo, Long> {
 }
